requirements.txt
django
pillow
whitenoise
djangorestframework
mediapipe